requirements.txt
streamlit
Pillow
imagehash
pandas
